requirements.txt
asgiref==3.2.7
asn1crypto==1.0.1
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
crypto==1.4.1
cryptography==2.7
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.1.1
django-phone-field==1.8.0
idna==2.9
Naked==0.1.31
oauthlib==3.1.0
Pillow==7.1.1
psycopg2==2.8.5
pycryptodome==3.9.7
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
shellescape==3.8.1
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.9
pycrypto==2.6.1
pycryptodome==3.9.7
gunicorn==20.0.4
django_heroku==0.3.1
